refactored models.py to be split accordingly to a models dir
