+++ nhl_api_py/core/api.py
@@ -7,8 +7,9 @@ from requests import request
 
 from nhl_api_py.core.decorators import timing
 from nhl_api_py.core.error_exceptions import ResponseError
-from nhl_api_py.core.models import Boxscore, Game, Play, Team
 from nhl_api_py.core.response import Response
+from nhl_api_py.models.game import Boxscore, Game, Play
+from nhl_api_py.models.team import Team
 
 logger = logging.getLogger(__name__)
